requirements.txt
PyQt6==6.6.1
qrcode[pil]==7.4.2
requests==2.31.0
Pillow==10.1.0
